refactor: Replace fixture
